+++ Source/Core/RequestState.js
@@ -6,16 +6,58 @@ define([
     'use strict';
 
     /**
-     * @private
+     * State of the request.
+     *
+     * @exports RequestState
      */
     var RequestState = {
-        UNISSUED : 0,   // Initial unissued state.
-        ISSUED : 1,     // Issued but not yet active. Will become active when open slots are available.
-        ACTIVE : 2,     // Actual http request has been sent.
-        RECEIVED : 3,   // Request completed successfully.
-        CANCELLED : 4,  // Request was cancelled, either explicitly or automatically because of low priority.
-        FAILED : 5,     // Request failed.
-        IGNORED : 6     // For RequestScheduler.clearForSpecs - lets requests finish but doesn't contribute to statistics.
+        /**
+         * Initial unissued state.
+         *
+         * @type Number
+         * @constant
+         */
+        UNISSUED : 0,
+
+        /**
+         * Issued but not yet active. Will become active when open slots are available.
+         *
+         * @type Number
+         * @constant
+         */
+        ISSUED : 1,
+
+        /**
+         * Actual http request has been sent.
+         *
+         * @type Number
+         * @constant
+         */
+        ACTIVE : 2,
+
+        /**
+         * Request completed successfully.
+         *
+         * @type Number
+         * @constant
+         */
+        RECEIVED : 3,
+
+        /**
+         * Request was cancelled, either explicitly or automatically because of low priority.
+         *
+         * @type Number
+         * @constant
+         */
+        CANCELLED : 4,
+
+        /**
+         * Request failed.
+         *
+         * @type Number
+         * @constant
+         */
+        FAILED : 5
     };
 
     return freezeObject(RequestState);