+++ Source/Core/loadBlob.js
@@ -15,7 +15,7 @@ define([
      *
      * @param {String} url The URL of the data.
      * @param {Object} [headers] HTTP headers to send with the requests.
-     * @param {Request} [request] The request object.
+     * @param {Request} [request] The request object. Intended for internal use only.
      * @returns {Promise.<Blob>|undefined} a promise that will resolve to the requested data when loaded. Returns undefined if <code>request.throttle</code> is true and the request does not have high enough priority.
      *
      * @example