+++ Source/Core/loadCRN.js
@@ -28,7 +28,7 @@ define([
      *
      * @param {String|ArrayBuffer} urlOrBuffer The URL of the binary data or an ArrayBuffer.
      * @param {Object} [headers] HTTP headers to send with the requests.
-     * @param {Request} [request] The request object.
+     * @param {Request} [request] The request object. Intended for internal use only.
      * @returns {Promise.<CompressedTextureBuffer>|undefined} A promise that will resolve to the requested data when loaded. Returns undefined if <code>request.throttle</code> is true and the request does not have high enough priority.
      *
      * @exception {RuntimeError} Unsupported compressed format.